Reference Rectangles for character positioning
Created reference rectangles and put them in key parts of the screen, offscreen to the left, far left, left, center, right, far right, off right. Also added a script to get the horizontal position of these reference points, so other scripts in the future can get these locations easily.

use refrectpos.center_hpos to get the value for the horizontal center of the screen.

current problem: i made the project a hard 16:9 and some shit needs to be changed around to allow for that to vary again like we want. the script in place accounts for this like i didnt cap it, i just couldnt figure out how to make the rectagles automatically warp around with the window.